+++ commands/fs/delete-file.js
