+++ commands/fs/rename-file.js
