+++ commands/fs/show-file-content.js
