+++ commands/zip/compress.js
