+++ commands/zip/decompress.js
